pyproject.toml
[project]
name = "agent-iot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.4",
    "uvicorn>=0.35.0",
]
